[refactor]: rename point_io module class file names snake_case
