+++ spatialetl/point/io/ascii/xy_coordinates_reader.py
@@ -25,7 +25,7 @@ from __future__ import division, print_function, absolute_import
 import numpy as np
 import pandas
 
-from spatialetl.point.io.MultiPointReader import MultiPointReader
+from spatialetl.point.io.multi_point_reader import MultiPointReader
 
 
 class XYCoordinatesReader(MultiPointReader):
