+++ spatialetl/providers/common/netcdf/spatialetl/point/io/netcdf/DefaultWriter.py
@@ -30,7 +30,7 @@ from numpy import float32, float64, int32
 from spatialetl.point.LevelMultiPoint import LevelMultiPoint
 from spatialetl.point.TimeLevelMultiPoint import TimeLevelMultiPoint
 from spatialetl.point.TimeMultiPoint import TimeMultiPoint
-from spatialetl.point.io.MultiPointWriter import MultiPointWriter
+from spatialetl.point.io.multi_point_writer import MultiPointWriter
 from spatialetl.utils.VariableDefinition import VariableDefinition
 from spatialetl.utils.logger import logging
